// app/src/generated/models/Ben_ltkissuesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Ben_ltkissuesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Ben_ltkissuesstatecode = keyof typeof Ben_ltkissuesstatecode;
export const Ben_ltkissuesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Ben_ltkissuesstatuscode = keyof typeof Ben_ltkissuesstatuscode;

export interface Ben_ltkissuesBase {
  ben_area?: string;
  ben_context?: string;
  ben_description?: string;
  "ben_DuplicateOf@odata.bind"?: string;
  ben_kind?: string;
  ben_ltkissueid: string;
  ben_name?: string;
  ben_priority?: number;
  ben_reporteremail?: string;
  ben_reportername?: string;
  ben_resolution?: string;
  ben_status?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Ben_ltkissuesstatecode;
  statuscode?: Ben_ltkissuesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Ben_ltkissues extends Ben_ltkissuesBase {
  ben_duplicateofname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  organizationidname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  ben_duplicateof?: object;
  _ben_duplicateof_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  organizationid?: object;
  _organizationid_value?: string;
}